Split out payload filtering.
